Exercise 4.21.

+++ part4/bloglist/controllers/blogs.js
@@ -44,8 +44,32 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+
+    if (!blog) {
+      console.log("Blog not found with ID:", request.params.id);
+      return response.status(404).json({ error: "blog not found" });
+    }
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response
+        .status(403)
+        .json({ error: "not authorized to delete this blog" });
+    }
+
+    await Blog.findByIdAndDelete(request.params.id);
+
+    response.status(204).end();
+  } catch (error) {
+    response.status(401).json({ error: "token invalid" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
